requirements.txt
aniso8601==7.0.0
ariadne==0.12.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-GraphQL==2.0.1
graphene==2.1.8
graphene-file-upload==1.2.2
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.2.0
promise==2.3
python-multipart==0.0.5
requests==2.24.0
Rx==1.6.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
starlette==0.13.7
threadpoolctl==2.1.0
typing==3.7.4.3
typing-extensions==3.7.4.2
urllib3==1.25.10
Werkzeug==1.0.1
